pyproject.toml
[project]
name = "flet-spendings"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "faker>=37.4.2",
  "flet==0.28.2",
  "flet-webview>=0.1.0",
  "icecream>=2.1.5",
  "loguru>=0.7.3",
  "prettyformatter>=2.0.13",
  "pytest>=8.4.1",
  "python-dotenv>=1.1.1",
  "supabase>=2.17.0",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "flet-spendings"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

# Include the .env file in the build
include = [".env"]

[tool.flet.app]
path = "src"


[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.2",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.2"}
